Update changelog 2025.4.100 (#7115)
Also updated `README.md`. There were a few content changes listed below, but most of the changes are formatting changes to get it closer to the extension README. At worst this makes them easier to compare and notice missing content in the pylance-release README. But perhaps we can get to the point where updating is automated.

- Updated `python.analysis.languageServerMode` table to correct `python.analysis.supportRestructuredText` default and add line for `python.analysis.nodeExecutable`.
- Corrected typo in name of `python.analysis.includeAliasesFromUserFiles`
- Remove `python.analysis.enableNotebookDataTips` which no longer exists.
- Added `documentation`, `overridden`, and `callable` to semantic token modifier list.